requirements.txt
dj-database-url==0.4.1
Django==1.11.1
djangorestframework==3.6.3
gunicorn==19.6.0
psycopg2==2.6.2
pytz==2017.2
whitenoise==3.2
